+++ tests/test_create_new_account.py
@@ -2,22 +2,18 @@ import pytest
 import allure
 
 
-@allure.feature('Test ui')
 @allure.story('Test create account')
 @allure.title('Test create new account with all correct fields')
-@allure.severity(allure.severity_level.CRITICAL)
-def test_create_new_account_with_all_correct_fields(create_account, start_end):
+def test_create_new_account_with_all_correct_fields(create_account, start_end, set_allure):
     create_account.open_page()
     create_account.fill_all_form()
     create_account.check_alert_text('Thank you for registering with Main Website Store.')
 
 
-@allure.feature('Test ui')
 @allure.story('Test create account with old account')
 @allure.title('Test create create account with old account')
-@allure.severity(allure.severity_level.CRITICAL)
 @pytest.mark.regression
-def test_create_account_with_old_account(create_account, create_new_account, start_end):
+def test_create_account_with_old_account(create_account, create_new_account, start_end, set_allure):
     old_data = create_new_account
     create_account.open_page(delete_cookies='Yes')
     create_account.fill_all_form(old_data)
@@ -27,12 +23,11 @@ def test_create_account_with_old_account(create_account, create_new_account, sta
     )
 
 
-@allure.feature('Test ui')
 @allure.story('Test create account with empty all fields')
 @allure.title('Test create account with empty all fields')
 @allure.severity(allure.severity_level.NORMAL)
 @pytest.mark.smoke
-def test_create_account_with_empty_all_fields(create_account, start_end):
+def test_create_account_with_empty_all_fields(create_account, start_end, set_allure):
     required_field = 'This is a required field.'
     create_account.open_page()
     create_account.click_button_create_account()
